+++ changelog.md
@@ -271,6 +271,9 @@
 - Added `genasts.genAst` that avoids the problems inherent with `quote do` and can
   be used as a replacement.
 
+- Added `reflection` module.
+
+
 ## Language changes
 
 - `nimscript` now handles `except Exception as e`.